> > > >It appears that Ray has found the problem.  It looks like config.h should be > >included before <string.h> in the cc source.  In this case things compile > >OK.  Does this seem like a reasonable fix to you? > >       It seems strange. I wonder what changes? I wouldn't expect config.h >       to define any of the #ifdefs of string.h, i.e.: > > 118     #if defined __USE_SVID || defined __USE_BSD || defined  __USE_XOPEN_EX TENDED > 119     /* Duplicate S, returning an identical malloc'd string.  */ > 120     extern char *strdup (__const char *__s) __THROW __attribute_malloc__; > 121     #endif > >       I don't think Ray's change hurts anything, but I'm at a loss to >       explain why it fixes anything. > I think this is the problem (Ray please correct me if I'm wrong).
HAVE_STRDUP is defined in config.h

[aips2mgr@titan161 glish]$ grep STRDUP config.h
#define HAVE_STRDUP 1

However, because config.h doesn't get included until after Dict.h (which
includes List.h which includes Object.h which includes Str.h), when Str.h is
included, HAVE_STRDUP still hasn't been defined.  Thus the #ifndef
HAVE_STRDUP bit tries to compile.  On all systems so far, that has been OK
because the strdup definition in Str.h and the one in <string.h> are the
same.  However, on the ia64 the system version of string.h defines strdup to
throw an exception which causes the two to be incompatible, and hence the
compile time error.  By putting config.h before Dict.h in the include path,
HAVE_STRDUP is defined by the time Str.h is included, and hence #ifndef
HaVE_STRDUP is false and the conflicting bit of code is ignored (which it
should be).